+++ src/test/java/de/chrgroth/generictypesystem/model/GenericStructureAttributeTest.java
@@ -91,7 +91,8 @@ public class GenericStructureAttributeTest {
     }
 
     private GenericAttribute attribute(GenericStructure structure, Long id, String name) {
-        GenericAttribute attribute = new GenericAttribute(id, name, DefaultGenericAttributeType.STRING, null, false, false, null, null, null, null, null, null, null, null, null);
+        GenericAttribute attribute = new GenericAttribute(id, name, DefaultGenericAttributeType.STRING, null, false, false, null, null, null, null, null, null, null, null, null,
+                null);
         structure.getAttributes().add(attribute);
         return attribute;
     }