+++ src/test/java/de/chrgroth/generictypesystem/validation/impl/DefaultValidationServiceTypeAttributeCriteriaTest.java
@@ -45,7 +45,9 @@ public class DefaultValidationServiceTypeAttributeCriteriaTest extends BaseValid
         service = new DefaultValidationService(unitsLookupTestHelper, null);
         type = new GenericType(0l, "testType", "testGroup", null, null, null, null);
         errorKeys = new ArrayList<>();
-        if (testType.isList()) {
+        if (testType.isEnum()) {
+            errorKeys.add(new ValidationError(ATTRIBUTE_NAME, DefaultValidationServiceMessageKey.TYPE_ATTRIBUTE_ENUM_VALUE_NOT_AVAILABLE));
+        } else if (testType.isList()) {
             errorKeys.add(new ValidationError(ATTRIBUTE_NAME, DefaultValidationServiceMessageKey.TYPE_ATTRIBUTE_LIST_VALUE_TYPE_MANDATORY));
         } else if (testType.isStructure()) {
             errorKeys.add(new ValidationError(ATTRIBUTE_NAME, DefaultValidationServiceMessageKey.TYPE_ATTRIBUTE_STRUCTURE_STRUCTURE_MANDATORY));
@@ -56,7 +58,7 @@ public class DefaultValidationServiceTypeAttributeCriteriaTest extends BaseValid
     public void minMaxTest() {
 
         // create attribute with min / max
-        createAttribute(null, false, false, null, 0.0d, 10.0d, null, null, null, null, null, null);
+        createAttribute(null, false, false, null, 0.0d, 10.0d, null, null, null, null, null, null, null);
 
         // define expected error message keys
         if (!testType.isMinMaxCapable()) {
@@ -74,7 +76,7 @@ public class DefaultValidationServiceTypeAttributeCriteriaTest extends BaseValid
 
         // create attribute with min / max and min = max
         clearAttributes();
-        createAttribute(null, false, false, null, 10.0d, 10.0d, null, null, null, null, null, null);
+        createAttribute(null, false, false, null, 10.0d, 10.0d, null, null, null, null, null, null, null);
 
         // define expected error message keys
         errorKeys.clear();
@@ -85,7 +87,7 @@ public class DefaultValidationServiceTypeAttributeCriteriaTest extends BaseValid
 
         // create attribute with min / max and min > max
         clearAttributes();
-        createAttribute(null, false, false, null, 15.0d, 10.0d, null, null, null, null, null, null);
+        createAttribute(null, false, false, null, 15.0d, 10.0d, null, null, null, null, null, null, null);
 
         // define expected error message keys
         errorKeys.clear();
@@ -99,7 +101,7 @@ public class DefaultValidationServiceTypeAttributeCriteriaTest extends BaseValid
     public void stepTest() {
 
         // create attribute with step
-        createAttribute(null, false, false, null, null, null, 1.0d, null, null, null, null, null);
+        createAttribute(null, false, false, null, null, null, 1.0d, null, null, null, null, null, null);
 
         // define expected error message keys
         if (!testType.isStepCapable()) {
@@ -116,7 +118,7 @@ public class DefaultValidationServiceTypeAttributeCriteriaTest extends BaseValid
 
         // create attribute with negative step
         clearAttributes();
-        createAttribute(null, false, false, null, null, null, -0.1d, null, null, null, null, null);
+        createAttribute(null, false, false, null, null, null, -0.1d, null, null, null, null, null, null);
 
         // define expected error message keys
         errorKeys.add(new ValidationError(ATTRIBUTE_NAME, DefaultValidationServiceMessageKey.TYPE_ATTRIBUTE_STEP_NEGATIVE));
@@ -129,7 +131,7 @@ public class DefaultValidationServiceTypeAttributeCriteriaTest extends BaseValid
     public void patternTest() {
 
         // create attribute with pattern
-        createAttribute(null, false, false, null, null, null, null, "test pattern", null, null, null, null);
+        createAttribute(null, false, false, null, null, null, null, "test pattern", null, null, null, null, null);
 
         // define expected error message keys
         if (!testType.isPatternCapable()) {
@@ -144,14 +146,14 @@ public class DefaultValidationServiceTypeAttributeCriteriaTest extends BaseValid
     public void valueProposalDependenciesTest() {
 
         // create attribute to depend upon
-        createAttribute(null, false, false, null, null, null, null, null, null, null, null, null);
+        createAttribute(null, false, false, null, null, null, null, null, null, null, null, null, null);
         attribute.setType(DefaultGenericAttributeType.STRING);
         attribute.setId(100l);
         attribute.setName("dependOnMe");
         GenericAttribute attributeToDependUpon = attribute;
 
         // create attribute with value proposal dependency
-        createAttribute(null, false, false, null, null, null, null, null, null, null, new HashSet<>(Arrays.asList(attributeToDependUpon.getId())), null);
+        createAttribute(null, false, false, null, null, null, null, null, null, null, new HashSet<>(Arrays.asList(attributeToDependUpon.getId())), null, null);
 
         // define expected error message keys
         if (!testType.isValueProposalDependenciesCapable()) {
@@ -168,7 +170,7 @@ public class DefaultValidationServiceTypeAttributeCriteriaTest extends BaseValid
 
         // value proposal dependency on non existing attribute
         clearAttributes();
-        createAttribute(null, false, false, null, null, null, null, null, null, null, new HashSet<>(Arrays.asList(123l)), null);
+        createAttribute(null, false, false, null, null, null, null, null, null, null, new HashSet<>(Arrays.asList(123l)), null, null);
 
         // define expected error message keys
         errorKeys.clear();
@@ -179,7 +181,7 @@ public class DefaultValidationServiceTypeAttributeCriteriaTest extends BaseValid
 
         // value proposal dependency on myself
         clearAttributes();
-        createAttribute(null, false, false, null, null, null, null, null, null, null, new HashSet<>(Arrays.asList(0l)), null);
+        createAttribute(null, false, false, null, null, null, null, null, null, null, new HashSet<>(Arrays.asList(0l)), null, null);
 
         // define expected error message keys
         errorKeys.clear();
@@ -193,7 +195,7 @@ public class DefaultValidationServiceTypeAttributeCriteriaTest extends BaseValid
     public void unitsTest() {
 
         // create attribute with base unit only
-        createAttribute(null, false, false, null, null, null, null, null, null, null, null, base(units()));
+        createAttribute(null, false, false, null, null, null, null, null, null, null, null, base(units()), null);
 
         // define expected error message keys
         if (!testType.isUnitCapable()) {
@@ -210,11 +212,51 @@ public class DefaultValidationServiceTypeAttributeCriteriaTest extends BaseValid
 
         // create invalid unit
         clearAttributes();
-        createAttribute(null, false, false, null, null, null, null, null, null, null, null, nonBase(units()));
+        createAttribute(null, false, false, null, null, null, null, null, null, null, null, nonBase(units()), null);
 
         // validate type
         errorKeys.clear();
         errorKeys.add(new ValidationError(ATTRIBUTE_NAME, DefaultValidationServiceMessageKey.TYPE_ATTRIBUTE_UNITS_INVALID, 0l));
+        validateType(errorKeys.toArray(new ValidationError[errorKeys.size()]));
+    }
+
+    @Test
+    public void enumValuesTest() {
+
+        // create attribute without enum values
+        createAttribute(null, false, false, null, null, null, null, null, null, null, null, null, null);
+
+        // validate type
+        validateType(errorKeys.toArray(new ValidationError[errorKeys.size()]));
+
+        // clear
+        clearAttributes();
+
+        // create attribute with empty enum values
+        createAttribute(null, false, false, null, null, null, null, null, null, null, null, null, new HashSet<>());
+
+        // validate type
+        validateType(errorKeys.toArray(new ValidationError[errorKeys.size()]));
+
+        // no further checks if not enum
+        if (!testType.isEnum()) {
+            return;
+        }
+
+        // clear
+        clearAttributes();
+        errorKeys.clear();
+
+        // create attribute with enum values
+        Set<String> enumValues = new HashSet<>();
+        enumValues.add("foo");
+        createAttribute(null, false, false, null, null, null, null, null, null, null, null, null, enumValues);
+
+        if (!testType.isEnum()) {
+            errorKeys.add(new ValidationError(ATTRIBUTE_NAME, DefaultValidationServiceMessageKey.TYPE_ATTRIBUTE_ENUM_VALUE_NOT_ALLOWED));
+        }
+
+        // validate type
         validateType(errorKeys.toArray(new ValidationError[errorKeys.size()]));
     }
 
@@ -234,9 +276,9 @@ public class DefaultValidationServiceTypeAttributeCriteriaTest extends BaseValid
     }
 
     public void createAttribute(DefaultGenericAttributeType valueType, boolean unique, boolean mandatory, GenericStructure structure, Double min, Double max, Double step,
-            String pattern, GenericValue<?> defaultValue, String defaultValueCallback, Set<Long> valueProposalDependencies, GenericUnits units) {
+            String pattern, GenericValue<?> defaultValue, String defaultValueCallback, Set<Long> valueProposalDependencies, GenericUnits units, Set<String> enumValues) {
         attribute = new GenericAttribute(0l, ATTRIBUTE_NAME, testType, valueType, unique, mandatory, structure, min, max, step, pattern, defaultValue, defaultValueCallback,
-                valueProposalDependencies, units != null ? units.getId() : null);
+                valueProposalDependencies, units != null ? units.getId() : null, enumValues);
         if (units != null) {
             unitsLookupTestHelper.register(units);
         }