+++ RELEASE-NOTES.md
@@ -5,6 +5,7 @@ Release Notes
 - refactored units to be a standalone top-level type
 - added conversion function for unit based values
 - refactored persistence service interface and abstract persistence service base implementation (breaking API changes)
+- added attribute based enum definitions
 - ...
 
 # 0.4.0